More library cleanup
Using built in motor models and made ones for the encoders we have used or are currently using. Also using built in clock, has almost every method in the masquerade one minus one which is just one statement. This should make the actual coding process easier as we can depend on built in methods that we know should work and not have to define motor models in code and config

Signed-off-by: Keval Kataria <kevalkataria@gmail.com>