Improved servo programmer and added code for distance sensor
The servo values for the servo programmer is now stored in an array and the class is cleaned up more. I added the code for the distance sensor but need to determine what distance correlates to the number of rings in the hopper. To calibrate it I made a distance sensor tester teleop which will just output the raw values.